detection/HiFi_IFDL-main/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
einops==0.8.1
ffmpeg==1.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
kmeans-pytorch==0.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
supervision==0.25.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yacs==0.1.8
